display mod refactored (I will refactor events later)
